# approval-api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PeopleCore 전자결재 API",
    "description": "전자결재 시스템 (Approval) API 명세서 - 53개 엔드포인트",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    }
  ],
...
  "paths": {
    "/approval/document": {
      "post": {
        "tags": ["결재문서"],
        "summary": "문서 기안 (결재 요청)",
        "description": "문서 생성 + 결재선 + 채번을 한번에 처리",
        "operationId": "createDocument",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/X-User-Name" },
          { "$ref": "#/components/parameters/X-User-Department" },
          { "$ref": "#/components/parameters/X-User-Grade" },
          { "$ref": "#/components/parameters/X-User-Title" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentCreateRequest" },
              "example": {
                "formId": 1,
                "docTitle": "2026년 상반기 개발팀 채용요청",
                "docType": "기안",
                "docData": "{\"field1\":\"value1\",\"field2\":\"value2\"}",
                "isEmergency": false,
                "approvalLines": [
                  {
                    "empId": 2,
                    "empName": "강희계",
                    "empDeptName": "인사팀",
                    "empGrade": "부장",
                    "empTitle": "팀장",
                    "approvalRole": "결재자",
                    "lineStep": 1
                  },
                  {
                    "empId": 3,
                    "empName": "박경영",
                    "empDeptName": "경영지원팀",
                    "empGrade": "이사",
                    "empTitle": "본부장",
                    "approvalRole": "결재자",
                    "lineStep": 2
                  },
                  {
                    "empId": 4,
                    "empName": "김참조",
                    "empDeptName": "경영지원팀",
                    "empGrade": "과장",
                    "empTitle": null,
                    "approvalRole": "참조",
                    "lineStep": 0
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "문서 생성 성공",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" },
                "example": 1
              }
            }
          }
        }
      }
    },
    "/approval/document/{docId}": {
      "get": {
        "tags": ["결재문서"],
        "summary": "문서 상세 조회",
        "operationId": "getDocumentDetail",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/PathDocId" }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentDetailResponse" },
                "example": {
                  "docId": 1,
                  "docNum": "인사-채용-260402-001",
                  "docTitle": "2026년 상반기 개발팀 채용요청",
                  "docType": "기안",
                  "docData": "{\"field1\":\"value1\"}",
                  "approvalStatus": "진행중",
                  "isEmergency": false,
                  "docSubmittedAt": "2026-04-02T10:00:00",
                  "docCompleteAt": null,
                  "empId": 1,
                  "empName": "김인재",
                  "empDeptName": "경영",
                  "empGrade": "차장",
                  "empTitle": "차장",
                  "formHtml": "<div class='form-container'>...</div>",
                  "formName": "채용요청",
                  "approvalLines": [
                    {
                      "lineId": 1,
                      "empId": 2,
                      "empName": "강희계",
                      "empDeptName": "인사팀",
                      "empGrade": "부장",
                      "empTitle": "팀장",
                      "approvalRole": "결재자",
                      "lineStep": 1,
                      "approvalLineStatus": "승인",
                      "lineProcessedAt": "2026-04-02T11:00:00",
                      "lineRejectReason": null,
                      "isDelegated": false,
                      "isRead": true
                    },
                    {
                      "lineId": 2,
                      "empId": 3,
                      "empName": "박경영",
                      "empDeptName": "경영지원팀",
                      "empGrade": "이사",
                      "empTitle": "본부장",
                      "approvalRole": "결재자",
                      "lineStep": 2,
                      "approvalLineStatus": "대기",
                      "lineProcessedAt": null,
                      "lineRejectReason": null,
                      "isDelegated": false,
                      "isRead": false
                    }
                  ],
                  "attachments": [
                    { "attachId": 1, "fileName": "채용계획서.pdf", "fileSize": 1024000 },
                    { "attachId": 2, "fileName": "인원현황.xlsx", "fileSize": 512000 }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["결재문서"],
        "summary": "문서 수정",
        "operationId": "updateDocument",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/PathDocId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentUpdateRequest" },
              "example": {
                "docTitle": "2026년 상반기 개발팀 채용요청 (수정)",
                "docData": "{\"field1\":\"updated_value\"}",
                "isEmergency": true,
                "approvalLines": [
                  {
                    "empId": 2,
                    "empName": "강희계",
                    "empDeptName": "인사팀",
                    "empGrade": "부장",
                    "empTitle": "팀장",
                    "approvalRole": "결재자",
                    "lineStep": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "수정 성공" }
        }
      },
      "delete": {
        "tags": ["결재문서"],
        "summary": "임시저장 문서 삭제",
        "operationId": "deleteDocument",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/PathDocId" }
        ],
        "responses": {
          "200": { "description": "삭제 성공" }
        }
      }
    },
    "/approval/document/temp": {
      "post": {
        "tags": ["결재문서"],
        "summary": "임시 저장",
        "operationId": "saveTempDocument",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/X-User-Name" },
          { "$ref": "#/components/parameters/X-User-Department" },
          { "$ref": "#/components/parameters/X-User-Grade" },
          { "$ref": "#/components/parameters/X-User-Title" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentCreateRequest" },
              "example": {
                "formId": 1,
                "docTitle": "임시저장 문서",
                "docType": "기안",
                "docData": "{\"field1\":\"draft_value\"}",
                "isEmergency": false,
                "approvalLines": [
                  {
                    "empId": 2,
                    "empName": "강희계",
                    "empDeptName": "인사팀",
                    "empGrade": "부장",
                    "empTitle": "팀장",
                    "approvalRole": "결재자",
                    "lineStep": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "임시저장 성공",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" },
                "example": 5
              }
            }
          }
        }
      }
    },
    "/approval/document/temp/{docId}": {
      "put": {
        "tags": ["결재문서"],
        "summary": "임시저장 수정",
        "operationId": "updateTempDocument",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/PathDocId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentUpdateRequest" },
              "example": {
                "docTitle": "임시저장 수정된 제목",
                "docData": "{\"field1\":\"updated_draft\"}",
                "isEmergency": false,
                "approvalLines": [
                  {
                    "empId": 2,
                    "empName": "강희계",
                    "empDeptName": "인사팀",
                    "empGrade": "부장",
                    "empTitle": "팀장",
                    "approvalRole": "결재자",
                    "lineStep": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "수정 성공" }
        }
      }
    },
    "/approval/document/{docId}/submit": {
      "post": {
        "tags": ["결재문서"],
        "summary": "임시저장 -> 결재 요청 전환",
        "operationId": "submitDocument",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/X-User-Department" },
          { "$ref": "#/components/parameters/PathDocId" }
        ],
        "responses": {
          "200": { "description": "전환 성공" }
        }
      }
    },
    "/approval/document/{docId}/resubmit": {
      "post": {
        "tags": ["결재문서"],
        "summary": "반려 문서 재기안",
        "description": "REJECTED -> PENDING (수정 + 새 채번 + 결재선 초기화)",
        "operationId": "resubmitDocument",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/X-User-Department" },
          { "$ref": "#/components/parameters/PathDocId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentUpdateRequest" },
              "example": {
                "docTitle": "2026년 상반기 개발팀 채용요청 (재기안)",
                "docData": "{\"field1\":\"revised_value\"}",
                "isEmergency": false,
                "approvalLines": [
                  {
                    "empId": 2,
                    "empName": "강희계",
                    "empDeptName": "인사팀",
                    "empGrade": "부장",
                    "empTitle": "팀장",
                    "approvalRole": "결재자",
                    "lineStep": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "재기안 성공" }
        }
      }
    },
    "/approval/document/{docId}/recall": {
      "post": {
        "tags": ["결재문서"],
        "summary": "상신 취소 (회수)",
        "description": "PENDING -> CANCELED",
        "operationId": "recallDocument",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/PathDocId" }
        ],
        "responses": {
          "200": { "description": "회수 성공" }
        }
      }
    },
...
    "/approval/documents/temp": {
      "get": {
        "tags": ["문서목록-개인"],
        "summary": "임시저장 문서",
        "operationId": "getTempDocuments",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Id" },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/SearchStartDate" },
          { "$ref": "#/components/parameters/SearchEndDate" },
          { "$ref": "#/components/parameters/SearchFormId" },
          { "$ref": "#/components/parameters/SearchStatus" },
          { "$ref": "#/components/parameters/PageNumber" },
          { "$ref": "#/components/parameters/PageSize" },
          { "$ref": "#/components/parameters/PageSort" }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PageDocumentListResponseDto" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "X-User-Company": {
        "name": "X-User-Company",
        "in": "header",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "회사 ID (UUID)",
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
      "X-User-Id": {
        "name": "X-User-Id",
        "in": "header",
        "required": true,
        "schema": { "type": "integer", "format": "int64" },
        "description": "사원 ID",
        "example": 1
      },
      "X-User-Name": {
        "name": "X-User-Name",
        "in": "header",
        "required": true,
        "schema": { "type": "string" },
        "description": "사원 이름",
        "example": "김인재"
      },
      "X-User-Department": {
        "name": "X-User-Department",
        "in": "header",
        "required": true,
        "schema": { "type": "integer", "format": "int64" },
        "description": "부서 ID",
        "example": 10
      },
      "X-User-Grade": {
        "name": "X-User-Grade",
        "in": "header",
        "required": true,
        "schema": { "type": "string" },
        "description": "직급 ID",
        "example": "차장"
      },
      "X-User-Title": {
        "name": "X-User-Title",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "직책 ID (nullable)",
        "example": "팀장"
      },
      "PathDocId": {
        "name": "docId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" },
        "description": "문서 ID",
        "example": 1
      },
...
      "SearchQuery": {
        "name": "search",
        "in": "query",
        "required": false,
        "schema": { "type": "string" },
        "description": "키워드 검색",
        "example": "채용"
      },
      "SearchStartDate": {
        "name": "startDate",
        "in": "query",
        "required": false,
        "schema": { "type": "string", "format": "date" },
        "description": "검색 시작일 (yyyy-MM-dd)",
        "example": "2026-01-01"
      },
      "SearchEndDate": {
        "name": "endDate",
        "in": "query",
        "required": false,
        "schema": { "type": "string", "format": "date" },
        "description": "검색 종료일 (yyyy-MM-dd)",
        "example": "2026-12-31"
      },
      "SearchFormId": {
        "name": "formId",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "format": "int64" },
        "description": "양식 필터",
        "example": 1
      },
      "SearchStatus": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": { "type": "string" },
        "description": "상태 필터",
        "example": "진행중"
      },
      "PageNumber": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "default": 0 },
        "description": "페이지 번호 (0부터 시작)",
        "example": 0
      },
      "PageSize": {
        "name": "size",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "default": 10 },
        "description": "페이지 크기",
        "example": 10
      },
      "PageSort": {
        "name": "sort",
        "in": "query",
        "required": false,
        "schema": { "type": "string" },
        "description": "정렬 (예: createdAt,desc)",
        "example": "createdAt,desc"
      }
    },
    "schemas": {
      "ApprovalLineRequest": {
        "type": "object",
        "properties": {
          "empId": { "type": "integer", "format": "int64", "description": "결재자 사원 ID" },
          "empName": { "type": "string", "description": "결재자 이름" },
          "empDeptName": { "type": "string", "description": "결재자 부서명" },
          "empGrade": { "type": "string", "description": "결재자 직급" },
          "empTitle": { "type": "string", "description": "결재자 직책" },
          "approvalRole": { "type": "string", "description": "결재 역할 (결재자/합의자/참조/열람)" },
          "lineStep": { "type": "integer", "description": "결재 순서" }
        }
      },
      "DocumentCreateRequest": {
        "type": "object",
        "required": ["formId", "docTitle"],
        "properties": {
          "formId": { "type": "integer", "format": "int64", "description": "양식 ID" },
          "docTitle": { "type": "string", "description": "문서 제목" },
          "docType": { "type": "string", "description": "문서 유형" },
          "docData": { "type": "string", "description": "양식 데이터 (HTML/JSON string)" },
          "isEmergency": { "type": "boolean", "description": "긴급 여부", "default": false },
          "approvalLines": { "type": "array", "items": { "$ref": "#/components/schemas/ApprovalLineRequest" }, "description": "결재선 목록" }
        }
      },
      "DocumentUpdateRequest": {
        "type": "object",
        "properties": {
          "docTitle": { "type": "string", "description": "문서 제목" },
          "docData": { "type": "string", "description": "양식 데이터" },
          "isEmergency": { "type": "boolean", "description": "긴급 여부" },
          "approvalLines": { "type": "array", "items": { "$ref": "#/components/schemas/ApprovalLineRequest" }, "description": "결재선 목록" }
        }
      },
      "ApprovalLineResponse": {
        "type": "object",
        "properties": {
          "lineId": { "type": "integer", "format": "int64" },
          "empId": { "type": "integer", "format": "int64" },
          "empName": { "type": "string" },
          "empDeptName": { "type": "string" },
          "empGrade": { "type": "string" },
          "empTitle": { "type": "string" },
          "approvalRole": { "type": "string" },
          "lineStep": { "type": "integer" },
          "approvalLineStatus": { "type": "string", "description": "대기/승인/반려 등" },
          "lineProcessedAt": { "type": "string", "format": "date-time" },
          "lineRejectReason": { "type": "string" },
          "isDelegated": { "type": "boolean" },
          "isRead": { "type": "boolean" }
        }
      },
      "DocumentDetailResponse": {
        "type": "object",
        "properties": {
          "docId": { "type": "integer", "format": "int64" },
          "docNum": { "type": "string", "description": "문서번호 (예: 인사-채용-260402-001)" },
          "docTitle": { "type": "string" },
          "docType": { "type": "string" },
          "docData": { "type": "string" },
          "approvalStatus": { "type": "string", "description": "임시저장/진행중/완료/반려/회수" },
          "isEmergency": { "type": "boolean" },
          "docSubmittedAt": { "type": "string", "format": "date-time" },
          "docCompleteAt": { "type": "string", "format": "date-time" },
          "empId": { "type": "integer", "format": "int64", "description": "기안자 사원 ID" },
          "empName": { "type": "string", "description": "기안자 이름" },
          "empDeptName": { "type": "string", "description": "기안자 부서명" },
          "empGrade": { "type": "string", "description": "기안자 직급" },
          "empTitle": { "type": "string", "description": "기안자 직책" },
          "formHtml": { "type": "string", "description": "양식 HTML" },
          "formName": { "type": "string", "description": "양식명" },
          "approvalLines": { "type": "array", "items": { "$ref": "#/components/schemas/ApprovalLineResponse" } },
          "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/AttachmentListResponse" } }
        }
      },
      "DocumentListResponseDto": {
        "type": "object",
        "properties": {
          "docId": { "type": "integer", "format": "int64" },
          "docTitle": { "type": "string" },
          "docNum": { "type": "string" },
          "docStatus": { "type": "string" },
          "isEmergency": { "type": "boolean" },
          "formName": { "type": "string" },
          "drafterName": { "type": "string" },
          "drafterDept": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "hasAttachment": { "type": "boolean" }
        }
      },
      "PageDocumentListResponseDto": {
        "type": "object",
        "properties": {
          "content": { "type": "array", "items": { "$ref": "#/components/schemas/DocumentListResponseDto" } },
          "pageable": { "type": "object" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer" },
          "size": { "type": "integer" },
          "number": { "type": "integer" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "empty": { "type": "boolean" }
        }
      },
...
      "AttachmentListResponse": {
        "type": "object",
        "properties": {
          "attachId": { "type": "integer", "format": "int64" },
          "fileName": { "type": "string" },
          "fileSize": { "type": "integer", "format": "int64" }
        }
      },
...
    }
  }
}